refactor:社会保険 標準報酬月額テーブルのファイル修正
